[build/soong] [SDCLANG] Support default block in SD LLVM config file
To better support the use case when some modules use SD LLVM while SDCLANG is
set to false globaly, a default block is required in the SD LLVM config file.
SDCLANG_PATH has to be set in the default block so that even if SDCLANG is
false, some modules can still use SD LLVM if needed.

Change-Id: I99fc864596b731a12b3f9a0ab5f47ce448753acb
Signed-off-by: Eliminater74 <eliminater74@gmail.com>